+++ server/api/event/views.py
@@ -49,15 +49,3 @@ def delete_event(request, event_id):
     else:
         return Response("Event is not private")
 
-@api_view(['POST'])
-def register(request):
-    if request.method == 'POST':
-        data = request.body
-        firstname = data.get('firstname')
-        lastname = data.get('lastname')
-        password = data.get('password')
-        email = data.get('email')
-        print(firstname, lastname, password, email)
-        return Response('User registered successfully')
-    else:
-        return Response({'error': 'Method Not Allowed'}, status=405)
